+++ src/Systemic/ControllerTrait.php
@@ -10,7 +10,9 @@ declare(strict_types=1);
 namespace DecodeLabs\Systemic;
 
 use DecodeLabs\Coercion;
+use DecodeLabs\Deliverance\DataReceiver;
 use DecodeLabs\Eventful\Dispatcher\Select as SelectDispatcher;
+use DecodeLabs\Exceptional;
 
 trait ControllerTrait
 {
@@ -34,36 +36,60 @@ trait ControllerTrait
         $streams = $this->manifold->getStreams();
 
 
+        // Prepare broker data
+        foreach ($command->getInputProviders() as $provider) {
+            $provider->setReadBlocking(false);
+        }
+
+
         // Input stream
         if (
-            isset($streams[0]) &&
+            isset($this->manifold->streams[0]) &&
             $input = $this->getInputStream()
         ) {
-            $this->dispatcher->bindStreamRead($input, function ($input) use ($streams) {
-                $streams[0]->write($input->read(2048));
+            $input->setReadBlocking(false);
+
+            $this->dispatcher->bindStreamRead($input, function ($input) {
+                $this->manifold->streams[0]->write($input->read(2048));
             });
         }
 
         // Output stream
         if (isset($streams[1])) {
-            $this->dispatcher->bindStreamRead($streams[1], function ($out) {
+            $this->dispatcher->bindStreamRead($streams[1], function ($out) use ($command) {
                 if (null === ($data = $out->readAll())) {
                     return;
                 }
 
                 $this->consumeOutput($data);
+
+                foreach ($command->getOutputReceivers() as $receiver) {
+                    if (!$receiver->isWritable()) {
+                        continue;
+                    }
+
+                    $this->writeToReceiver($receiver, $data);
+                }
             });
         }
 
 
         // Error
         if (isset($streams[2])) {
-            $this->dispatcher->bindStreamRead($streams[2], function ($err) {
+            $this->dispatcher->bindStreamRead($streams[2], function ($err) use ($command) {
                 if (null === ($data = $err->readAll())) {
                     return;
                 }
 
                 $this->consumeError($data);
+
+                foreach ($command->getErrorReceivers() as $receiver) {
+                    if (!$receiver->isWritable()) {
+                        continue;
+                    }
+
+                    $this->writeToReceiver($receiver, $data);
+                }
             });
         }
 
@@ -75,25 +101,37 @@ trait ControllerTrait
         }
 
         // Ticks
-        $this->dispatcher->setTickHandler(function () {
+        $this->dispatcher->setTickHandler(function () use ($command) {
             $status = $this->manifold->getStatus();
             $running = ($status['running'] ?? false);
 
+
             // Provide input
-            if (null !== ($data = $this->provideInput())) {
-                if (
-                    $running &&
-                    isset($this->manifold->streams[0])
-                ) {
-                    while (strlen($data) > 0) {
-                        $written = $this->manifold->streams[0]->write($data, 2048);
-                        $data = substr($data, $written);
+            if (
+                $running &&
+                isset($this->manifold->streams[0])
+            ) {
+                // Controller data iterator
+                $this->writeToReceiver(
+                    $this->manifold->streams[0],
+                    $this->provideInput()
+                );
+
+                // Command input providers
+                foreach ($command->getInputProviders() as $provider) {
+                    if (!$provider->isReadable()) {
+                        continue;
                     }
+
+                    $this->writeToReceiver(
+                        $this->manifold->streams[0],
+                        $provider->readAll()
+                    );
                 }
             }
 
 
-            /* @phpstan-ignore-next-line */
+            // Complete process
             if (!$running) {
                 $this->registerCompletion(Coercion::toInt($status['exitcode'] ?? 0));
                 return false;
@@ -102,15 +140,43 @@ trait ControllerTrait
 
         if (empty($streams)) {
             $this->dispatcher->bindTimer('keepAlive', 1, function () {
-                // Lah di dah
+                // We need a timer to keep the tick handler running if there are no streams
             });
         }
 
 
+        // Run dispatcher
         $this->dispatcher->listen();
+
+        // Shutdown
         $this->dispatcher->removeAllBindings();
         $this->manifold->close();
 
         return $process;
+    }
+
+    protected static function writeToReceiver(
+        DataReceiver $receiver,
+        ?string $data
+    ): void {
+        if ($data === null) {
+            return;
+        }
+
+        $error = 0;
+
+        while (strlen($data) > 0) {
+            $written = $receiver->write($data, 2048);
+
+            if ($written === 0) {
+                if (++$error > 5) {
+                    throw Exceptional::Runtime('Unable to write to data receiver');
+                }
+
+                usleep(10000);
+            }
+
+            $data = substr($data, $written);
+        }
     }
 }