+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added Deliverance Broker interfaces to Commands / Controllers
+
 ## v0.11.1 (2022-11-30)
 * Fixed run() calls over non TTY CLI sapis